Move applyReplaceUsingPlugin to the replacementVariableHelpers
