+++ js/src/containers/SnippetEditor.js
@@ -1,14 +1,8 @@
 import React from "react";
 import { connect } from "react-redux";
 import { SnippetEditor } from "@yoast/search-metadata-previews";
-import {
-	get,
-	identity,
-} from "lodash-es";
 import { __ } from "@wordpress/i18n";
 import { dispatch as wpDataDispatch } from "@wordpress/data";
-import analysis from "yoastseo";
-const { stripHTMLTags: stripFullTags } = analysis.string;
 
 import {
 	switchMode,
@@ -16,52 +10,7 @@ import {
 } from "../redux/actions/snippetEditor";
 import { updateAnalysisData } from "../redux/actions/analysisData";
 import SnippetPreviewSection from "../components/SnippetPreviewSection";
-
-/**
- * Runs the legacy replaceVariables function on the data in the snippet preview.
- *
- * @param {Object} data             The snippet preview data object.
- * @param {string} data.title       The snippet preview title.
- * @param {string} data.url         The snippet preview url: baseUrl with the slug.
- * @param {string} data.description The snippet preview description.
- *
- * @returns {Object} Returns the data object in which the placeholders have been replaced.
- */
-const legacyReplaceUsingPlugin = function( data ) {
-	const replaceVariables = get( window, [ "YoastSEO", "wp", "replaceVarsPlugin", "replaceVariables" ], identity );
-
-	return {
-		url: data.url,
-		title: stripFullTags( replaceVariables( data.title ) ),
-		description: stripFullTags( replaceVariables( data.description ) ),
-	};
-};
-
-/**
- * Apply replaceVariables function on the data in the snippet preview.
- *
- * @param {Object} data             The snippet preview data object.
- * @param {string} data.title       The snippet preview title.
- * @param {string} data.url         The snippet preview url: baseUrl with the slug.
- * @param {string} data.description The snippet preview description.
- *
- * @returns {Object} Returns the data object in which the placeholders have been replaced.
- */
-const applyReplaceUsingPlugin = function( data ) {
-	// If we do not have pluggable loaded, apply just our own replace variables.
-	const pluggable = get( window, [ "YoastSEO", "app", "pluggable" ], false );
-	if ( ! pluggable || ! get( window, [ "YoastSEO", "app", "pluggable", "loaded" ], false ) ) {
-		return legacyReplaceUsingPlugin( data );
-	}
-
-	const applyModifications = pluggable._applyModifications.bind( pluggable );
-
-	return {
-		url: data.url,
-		title: stripFullTags( applyModifications( "data_page_title", data.title ) ),
-		description: stripFullTags( applyModifications( "data_meta_desc", data.description ) ),
-	};
-};
+import { applyReplaceUsingPlugin } from "../helpers/replacementVariableHelpers";
 
 /**
  * Process the snippet editor form data before it's being displayed in the snippet preview.
